% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPWmeasures.R
\name{AIPWmeasures}
\alias{AIPWmeasures}
\title{Estimate standard measures of predictive accuracy for two-phase designs using AIPW}
\usage{
AIPWmeasures(time, event, X, subcohort, aug.weights.x = NULL, risk.threshold,
  landmark.time, weight.method = c("Aug", "True"), design = c("CCH", "NCC"),
  calculate.sd = TRUE, smoothing.par = 0.7, pnf.threshold = NULL,
  pcf.threshold = NULL, ncc.nmatch = NULL)
}
\arguments{
\item{time}{vector of time to event}

\item{event}{status indicator (1 for observed failure, 0 for censoring)}

\item{X}{matrix or data frame of covariates}

\item{subcohort}{indicator for selection into the subcohort (1 for selected, 0 for not selected)}

\item{aug.weights.x}{vector of variable to be used to derive augmented inverse probability weights.}

\item{risk.threshold}{vector of risk thresholds on the absolute risk scale used to calcuate cutoff-based summary measures.}

\item{landmark.time}{numeric value specifying landmark time at which to estimate the performance measures.}

\item{weight.method}{either "Aug" for augmented ipw or "True" for true ipw.}

\item{design}{either "CCH" for case-cohort or "NCC" for nested case control specifying the subcohort design used.}

\item{calculate.sd}{should analytic standard errors be calculated.}

\item{smoothing.par}{nearest neighbor smoothing parameter used in locfit (default = 0.7)}

\item{pnf.threshold}{numeric threshold value for which to calculate PNF(bb)}

\item{pcf.threshold}{numeric threshold value for which to calculate PCF(vv)}

\item{ncc.nmatch}{For design = "NCC", specify the number of controls matched per case.}
}
\value{
data.frame with estimates of AUC, IDI, ITPR, IFPR, TPR, FPR, PPV, NPV and net benefit (NB) and standard errors.
}
\description{
Estimate measures of predictive accuracy using augmented inverse probability weights (ipw) or true ipw for two-phase biomarker validation studies
}
\note{
variance calculations are unavailable for the NB measure.
}
\examples{

data(CCHsimdata)

predict.time <- 0.75

## augmented ipw
AIPWmeasures( time = CCHsimdata$xi, event = CCHsimdata$di,
             X = cbind(CCHsimdata$y1, CCHsimdata$y2),
            subcohort = CCHsimdata$vi,
             aug.weights.x =   CCHsimdata$y1,
             risk.threshold  =  c(.05, .3),
             landmark.time = predict.time,
             weight.method = 'Aug',
             design = "CCH",
             smoothing.par = 0.9,
             calculate.sd  = TRUE,
             pnf.threshold = 0.85,
             pcf.threshold = 0.8)


#simulated data from a ncc design with nmatch = 2
data("NCCsimdata")
AIPWmeasures( time = NCCsimdata$xi, event = NCCsimdata$di,
             X = cbind(NCCsimdata$y1, NCCsimdata$y2),
             subcohort = NCCsimdata$vi,
             aug.weights.x =   NCCsimdata$y1,
             risk.threshold  =  c(.01, .03),
             landmark.time = predict.time,
             weight.method = 'Aug',
             design = "NCC",
             smoothing.par = 0.9,
             calculate.sd  = TRUE,
             pnf.threshold = 0.85,
             pcf.threshold = 0.8,
             ncc.nmatch = 2)

}

